Refactor component to be used as View and ViewController
