+++ CybridSDK/Components/Trade/Modal/TradeSuccessModalView.swift
@@ -197,6 +197,6 @@ extension TradeSuccessModalView {
     let cryptoAmount: String
     let cryptoCode: String
     let transactionFee: String
-    let quoteType: _TradeType
+    let quoteType: TradeType
   }
 }
